Move Classic-only test helpers to tests/classic-fixtures.ts
`tests/helpers.ts` had three Classic-specific helpers (`createMockApi`,
`createPopulatedRegistry`, `assertDeviceType`) sitting alongside
scope-neutral helpers — an inconsistency that became visible after
splitting fixtures into `tests/classic-fixtures.ts` and
`tests/home-fixtures.ts` with explicit `classic*` / `home*` prefixes.

Migrate and rename for consistency with the project-wide Classic/Home
convention:

- `createMockApi` → `createMockClassicApi`
- `createPopulatedRegistry` → `populatedClassicRegistry`
- `assertDeviceType` → `assertClassicDeviceType`

`tests/helpers.ts` now contains only scope-neutral helpers
(`cast`, `mock`, `defined`, `matchObject`, `createLogger`,
`mockResponse`, `mockFetchResponse`, `createHttpError`,
`createServerError`, `createUnauthorizedError`, `createMockHttpClient`,
`createSettingStore`).

No behaviour change. Verification: lint, format, typecheck, vitest
(641 passed), 100% coverage.
