+++ tests/classic-fixtures.ts
@@ -1,8 +1,15 @@
+import { vi } from 'vitest'
+
+import type { ClassicAPIAdapter, SyncCallback } from '../src/api/index.ts'
 import {
   ClassicDeviceType,
   ClassicLabelType,
   ClassicOperationModeZone,
 } from '../src/constants.ts'
+import {
+  type ClassicDeviceAny,
+  ClassicRegistry,
+} from '../src/entities/index.ts'
 import {
   type ClassicAreaDataAny,
   type ClassicBuildingData,
@@ -294,3 +301,108 @@ export const classicRawDevice = (
     Type: 0,
     ...overrides,
   })
+
+// ---------------------------------------------------------------------------
+// Mock factories
+// ---------------------------------------------------------------------------
+
+const MOCK_SIGNAL_RSSI_DBM = -60
+
+export const createMockClassicApi = (
+  overrides: Partial<ClassicAPIAdapter> = {},
+): ClassicAPIAdapter =>
+  mock<ClassicAPIAdapter>({
+    fetch: vi.fn<ClassicAPIAdapter['fetch']>().mockResolvedValue([]),
+    getEnergy: vi.fn<ClassicAPIAdapter['getEnergy']>(),
+    getErrorEntries: vi.fn<ClassicAPIAdapter['getErrorEntries']>(),
+    getErrorLog: vi.fn<ClassicAPIAdapter['getErrorLog']>(),
+    getFrostProtection: vi
+      .fn<ClassicAPIAdapter['getFrostProtection']>()
+      .mockResolvedValue(cast({ data: { FPEnabled: false } })),
+    getGroup: vi.fn<ClassicAPIAdapter['getGroup']>(),
+    getHolidayMode: vi
+      .fn<ClassicAPIAdapter['getHolidayMode']>()
+      .mockResolvedValue(cast({ data: { HMEnabled: false } })),
+    getHourlyTemperatures: vi.fn<ClassicAPIAdapter['getHourlyTemperatures']>(),
+    getInternalTemperatures:
+      vi.fn<ClassicAPIAdapter['getInternalTemperatures']>(),
+    getOperationModes: vi.fn<ClassicAPIAdapter['getOperationModes']>(),
+    getSignal: vi.fn<ClassicAPIAdapter['getSignal']>().mockResolvedValue(
+      cast({
+        data: {
+          Data: [[{ Data: [MOCK_SIGNAL_RSSI_DBM], Name: 'ClassicDevice' }]],
+          Labels: ['12:00'],
+        },
+      }),
+    ),
+    getTemperatures: vi.fn<ClassicAPIAdapter['getTemperatures']>(),
+    getTiles: cast(
+      vi
+        .fn<ClassicAPIAdapter['getTiles']>()
+        .mockResolvedValue(cast({ data: {} })),
+    ),
+    getValues: vi.fn<ClassicAPIAdapter['getValues']>(),
+    notifySync: vi.fn<SyncCallback>().mockResolvedValue(),
+    updateFrostProtection: vi
+      .fn<ClassicAPIAdapter['updateFrostProtection']>()
+      .mockResolvedValue(cast({ data: { Success: true } })),
+    updateGroupState: vi.fn<ClassicAPIAdapter['updateGroupState']>(),
+    updateHolidayMode: vi
+      .fn<ClassicAPIAdapter['updateHolidayMode']>()
+      .mockResolvedValue(cast({ data: { Success: true } })),
+    updatePower: vi
+      .fn<ClassicAPIAdapter['updatePower']>()
+      .mockResolvedValue({ data: true }),
+    updateValues: cast(vi.fn<ClassicAPIAdapter['updateValues']>()),
+    ...overrides,
+  })
+
+// ---------------------------------------------------------------------------
+// Type guards & registry helpers
+// ---------------------------------------------------------------------------
+
+export function assertClassicDeviceType<T extends ClassicDeviceType>(
+  device: ClassicDeviceAny | undefined,
+  type: T,
+): asserts device is Extract<ClassicDeviceAny, { type: T }>
+export function assertClassicDeviceType(
+  device: ClassicDeviceAny | undefined,
+  type: ClassicDeviceType,
+): void {
+  if (device?.type !== type) {
+    throw new Error(
+      `Expected device of type ${String(type)}, got ${device ? String(device.type) : 'undefined'}`,
+    )
+  }
+}
+
+/**
+ * Build a `ClassicRegistry` populated with the provided hierarchy in a
+ * single call. Replaces the repeated 5-line
+ * `new ClassicRegistry() + syncBuildings + syncFloors + syncAreas + syncDevices`
+ * pattern found in multiple test files.
+ * @param data - Flat arrays of buildings, floors, areas, and devices.
+ * @param data.areas - ClassicArea rows to sync.
+ * @param data.buildings - ClassicBuilding rows to sync.
+ * @param data.devices - ClassicDevice rows to sync.
+ * @param data.floors - ClassicFloor rows to sync.
+ * @returns A fully synced `ClassicRegistry` instance.
+ */
+export const populatedClassicRegistry = ({
+  areas,
+  buildings,
+  devices,
+  floors,
+}: {
+  areas: ClassicAreaDataAny[]
+  buildings: ClassicBuildingData[]
+  devices: ClassicListDeviceAny[]
+  floors: ClassicFloorData[]
+}): ClassicRegistry => {
+  const registry = new ClassicRegistry()
+  registry.syncBuildings(buildings)
+  registry.syncFloors(floors)
+  registry.syncAreas(areas)
+  registry.syncDevices(devices)
+  return registry
+}
